Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int fl;
+    float fl;
     scanf("%f", &fl);
     printf("You entered: %.2f", fl);
     return 0;
